pull out delete submission into api
