+++ commands/db/host/deletesubmission.py
@@ -3,28 +3,13 @@ import os
 import discord
 from discord.ext import commands
 from dotenv import load_dotenv
-from sqlalchemy import select, delete
 
-from api.db_classes import Submissions, get_session, Tasks
+from api.submissions import delete_submission
 from api.utils import has_host_role
 from discord_ext.discord_utils import edit_submission_list
 
 load_dotenv()
 DEFAULT = os.getenv('DEFAULT')
-
-
-async def reorder_primary_keys():
-    async with get_session() as session:
-        # Retrieve the submissions
-        query = select(Submissions).order_by(Submissions.index)
-        result = await session.execute(query)
-        submissions = result.scalars().all()
-
-        # Reassign indexes
-        for idx, submission in enumerate(submissions, start=1):
-            submission.index = idx
-
-        await session.commit()
 
 
 class DeleteSubmission(commands.Cog):
@@ -34,30 +19,13 @@ class DeleteSubmission(commands.Cog):
     @commands.hybrid_command(name="delete-submission", description="Delete someone's submission", with_app_command=True)
     @has_host_role()
     async def command(self, ctx, user: discord.Member):
-        async with get_session() as session:
-            currently_running = (await session.execute(select(Tasks).where(Tasks.is_active == 1))).first()
-            if not currently_running:
-                return await ctx.send("There is no ongoing task")
+        # delete the user's submission
+        if await delete_submission(user.id, user.display_name):
+            # Update submission list
+            await edit_submission_list(self)
 
-        async with get_session() as session:
-            data = (await session.scalars(select(Submissions).where(Submissions.user_id == user.id))).first()
-
-        if data is None:
-            return await ctx.reply(f"{user} has no submission.")
-
-        # Delete submission from db
-        async with get_session() as session:
-            await session.execute(delete(Submissions).where(Submissions.user_id == user.id))
-            await session.commit()
-
-        # Re-arrange the indexes in the submission table so that they are no gaps between numbers
-        await reorder_primary_keys()
-
-        # Update submission list
-        await edit_submission_list(self)
-
-        await ctx.send(f"{user.display_name}'s submission has been deleted.",
-                       allowed_mentions=discord.AllowedMentions.none(), suppress_embeds=True)
+            await ctx.send(f"{user.display_name}'s submission has been deleted.",
+                           allowed_mentions=discord.AllowedMentions.none(), suppress_embeds=True)
 
 
 async def setup(bot) -> None:
